Create Parser.py
The parser works. The only thing we have to check is if I got the formatting right for the return message. From looking at the arduino code it looks like it pretty much just wants the same as the output string but im not sure. Also for the recipes, I'm not sure how to do it with xml files but it's really easy with just regular text files. Assuming everything works, the only thing left to do for this is slap it inside a while loop.